same as below for here.
2003-01-05  Manish Singh  <yosh@gimp.org>

        * plug-ins/script-fu/interp_regex.c: same as below for here.

+++ plug-ins/script-fu/siod/regex.c
@@ -141,8 +141,8 @@ regex_prin1 (LISP ptr, struct gen_printio *f)
   char buffer[256];
   regex_t *p;
   p = get_tc_regex (ptr)->r;
-  sprintf (buffer, "#<REGEX %p nsub=%lu",
-	   p, p->re_nsub);
+  sprintf (buffer, "#<REGEX %p nsub=%d",
+	   p, (int) p->re_nsub);
   gput_st (f, buffer);
   gput_st (f, ">");
 }
